Create Anagram.c
An anagram is a word or phrase that's formed by rearranging the letters of another word or phrase. For example, the letters that make up “A decimal point” can be turned into the anagram “I'm a dot in place.” People mainly make anagrams just for fun, but sometimes they're used as pseudonyms or codes.